fix: capability-driven split-A/V audio + Long clock-offset
Split-A/V to a receiver that can't render the source audio codec
(e.g. TrueHD-Atmos on a Dolby-Digital-Plus soundbar / Google TV
Streamer) played silently: the receiver's audio track was dropped
(groups=0) so the audioFormat "not supported" signal never fired and
the legacy direct-play codec table optimistically streamed it raw.

- Receiver advertises its authoritative passthrough-capable codecs
  (ReceiverAudioCodecs.fromCapabilities) on a new SpatialFin beacon
  extension PlaybackUpdateMessage.supportedAudioCodecs.
- castSpatialItemSplitAv now decides direct-stream vs transcode from
  that capability set (never a hardcoded table): direct-stream when
  the chain can render the codec (lossless/Atmos preserved), else a
  constrained-profile HLS transcode (JellyfinRepository
  .getAudioTranscodeStreamUrl) to the best codec the chain supports,
  unbounded bitrate so the only transcode reason is the codec.
- Forced transcode via enableDirectPlay/Stream=false (a secondary
  compatible audio track no longer defeats it) and the HLS path
  replaces the static=false&audioCodec= raw-/stream rewrite that
  produced an unreadable container (receiver groups=0).
- First cast to an unseen receiver self-corrects: the first
  authoritative beacon triggers a one-shot transparent re-cast.
- app/unified now depends on media3-exoplayer-hls; without it the
  receiver's DefaultMediaSourceFactory fell back to progressive and
  could not parse the .m3u8 transcode (groups=0). Receiver now maps
  .m3u8/.mpd to media3's canonical MIME before the container hint.
- ClockOffsetEstimator.offsetMs() / BeaconState.clockOffsetMs are now
  Long, not Int. θ is the gap between the two devices' elapsedRealtime
  epochs and exceeds 2^31 ms when uptimes differ by weeks; truncating
  wrapped a ~4.3e9 ms offset to ~8.3e7, corrupting expectedXrPositionMs
  into multi-gigasecond drift -> hard-seek every beacon -> degrade ->
  pause cascaded to the XR video master.

Verified live on Galaxy XR (video master) + Google TV Streamer
(audio, DD+/Atmos soundbar) with a TrueHD-Atmos source: audio plays
and video no longer self-pauses.

+++ app/unified/src/main/java/dev/spatialfin/fcast/session/SplitAvPolicy.kt
@@ -117,7 +117,7 @@ object SplitAvPolicy {
          * FCast v4: estimated receiver↔sender clock offset θ (`receiver ≈ sender + θ`) from
          * [ClockOffsetEstimator]. Null until θ converges / pre-v4 peer → legacy path.
          */
-        val clockOffsetMs: Int? = null,
+        val clockOffsetMs: Long? = null,
         /**
          * FCast v4: the receiver's monotonic clock when it sampled [beaconStreamPositionMs]
          * (`PlaybackUpdateMessage.monotonicSampleMs`). Null pre-v4 → legacy path.
